feat: 게시글 서비스 도메인 경로 변경 및 엔티티 수정
