--- article-service/src/main/java/com/newcord/articleservice/block/entity/BlockCreatedBy.java
@@ -1,11 +1,0 @@
-package com.newcord.articleservice.block.entity;
-
-
-import java.time.LocalDateTime;
-import lombok.Builder;
-
-@Builder
-public class BlockCreatedBy {
-    private String creator_id;
-    private LocalDateTime created_at;
-}
