--- article-service/src/main/java/com/newcord/articleservice/block/entity/BlockParent.java
@@ -1,9 +1,0 @@
-package com.newcord.articleservice.block.entity;
-
-import lombok.Builder;
-
-@Builder
-public class BlockParent {
-    private String type;
-    private String page_id;
-}
