--- article-service/src/main/java/com/newcord/articleservice/clip/entity/Clip.java
@@ -1,10 +1,0 @@
-package com.newcord.articleservice.clip.entity;
-
-import lombok.Builder;
-
-//TODO: Clip 내부 변수 기획 필요
-@Builder
-public class Clip {
-    private String id;
-
-}
