--- article-service/src/main/java/com/newcord/articleservice/clip/repository/ClipRepository.java
@@ -1,8 +1,0 @@
-package com.newcord.articleservice.clip.repository;
-
-import com.newcord.articleservice.clip.entity.Clip;
-import org.springframework.data.mongodb.repository.MongoRepository;
-
-public interface ClipRepository extends MongoRepository<Clip, String> {
-
-}
