+++ article-service/src/main/java/com/newcord/articleservice/domain/block/entity/BlockCreatedBy.java
@@ -1,0 +1,13 @@
+package com.newcord.articleservice.domain.block.entity;
+
+
+import java.time.LocalDateTime;
+import lombok.Builder;
+import org.springframework.data.annotation.CreatedDate;
+
+@Builder
+public class BlockCreatedBy {
+    private String creator_id;
+    @CreatedDate
+    private LocalDateTime created_at;
+}
