+++ article-service/src/main/java/com/newcord/articleservice/domain/block/entity/BlockParent.java
@@ -1,0 +1,9 @@
+package com.newcord.articleservice.domain.block.entity;
+
+import lombok.Builder;
+
+@Builder
+public class BlockParent {
+    private String type;
+    private String page_id;
+}
