+++ article-service/src/main/java/com/newcord/articleservice/domain/hashtags/entity/Hashtags.java
@@ -1,0 +1,24 @@
+package com.newcord.articleservice.domain.hashtags.entity;
+
+
+import com.newcord.articleservice.global.common.BaseTimeEntity;
+import jakarta.persistence.Entity;
+import jakarta.persistence.GeneratedValue;
+import jakarta.persistence.GenerationType;
+import jakarta.persistence.Id;
+import lombok.*;
+
+import java.io.Serializable;
+
+@Entity
+@Getter
+@NoArgsConstructor
+@AllArgsConstructor
+@Builder
+public class Hashtags extends BaseTimeEntity {
+    @Id
+    @GeneratedValue(strategy = GenerationType.IDENTITY)
+    private Integer id;
+    private String tagName;
+}
+
