+++ article-service/src/main/java/com/newcord/articleservice/domain/posts/entity/Posts.java
@@ -1,0 +1,27 @@
+package com.newcord.articleservice.domain.posts.entity;
+
+import com.newcord.articleservice.global.common.BaseTimeEntity;
+import jakarta.persistence.Entity;
+import jakarta.persistence.GeneratedValue;
+import jakarta.persistence.GenerationType;
+import jakarta.persistence.Id;
+import lombok.*;
+
+import java.io.Serializable;
+import java.sql.Timestamp;
+
+@Entity
+@Getter
+@NoArgsConstructor
+@AllArgsConstructor
+@Builder
+public class Posts extends BaseTimeEntity {
+    @Id
+    @GeneratedValue(strategy = GenerationType.IDENTITY)
+    private Integer id;
+    private String thumbnail;
+    private String title;
+    private String body;
+    private String status;
+    private Integer views;
+}
