--- article-service/src/main/java/com/newcord/articleservice/hashtags/entity/Hashtags.java
@@ -1,23 +1,0 @@
-package com.newcord.articleservice.hashtags.entity;
-
-
-import jakarta.persistence.Entity;
-import jakarta.persistence.GeneratedValue;
-import jakarta.persistence.GenerationType;
-import jakarta.persistence.Id;
-import lombok.*;
-
-import java.io.Serializable;
-
-@Entity
-@Getter
-@NoArgsConstructor
-@AllArgsConstructor
-@Builder
-public class Hashtags implements Serializable {
-    @Id
-    @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Integer id;
-    private String tagName;
-}
-
