--- article-service/src/main/java/com/newcord/articleservice/posts_hashtag/entity/PostsHashtag.java
@@ -1,20 +1,0 @@
-package com.newcord.articleservice.posts_hashtag.entity;
-
-import jakarta.persistence.*;
-import lombok.*;
-
-import java.io.Serializable;
-
-@Entity
-@Getter
-@NoArgsConstructor
-@AllArgsConstructor
-@Builder
-public class PostsHashtag implements Serializable {
-    @Id
-    @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Integer id;
-
-    private Integer post_id;
-    private Integer hashtag_id;
-}
